refactor(base): move amount to base mod
