+++ rust/src/base/mod.rs
@@ -1,0 +1,2 @@
+//! Indexer base types
+pub mod amount;
