--- rust/src/ledger/amount.rs
@@ -1,87 +1,0 @@
-use crate::{
-    constants::{MINA_SCALE, MINA_SCALE_DEC},
-    utility::functions::nanomina_to_mina,
-};
-use anyhow::anyhow;
-use rust_decimal::{prelude::ToPrimitive, Decimal};
-use serde::{Deserialize, Serialize};
-use std::ops::{Add, Sub};
-
-#[derive(
-    Debug, PartialEq, Eq, PartialOrd, Ord, Clone, Copy, Default, Hash, Serialize, Deserialize,
-)]
-pub struct Amount(pub u64);
-
-impl Amount {
-    pub fn new(amount: u64) -> Self {
-        Self(amount * MINA_SCALE)
-    }
-}
-
-// operations
-
-impl Add<Amount> for Amount {
-    type Output = Amount;
-
-    fn add(self, rhs: Amount) -> Self::Output {
-        Self(self.0 + rhs.0)
-    }
-}
-
-impl Add<u64> for Amount {
-    type Output = Amount;
-
-    fn add(self, rhs: u64) -> Self::Output {
-        Self(self.0 + rhs)
-    }
-}
-
-impl Add<i64> for Amount {
-    type Output = Amount;
-
-    fn add(self, rhs: i64) -> Self::Output {
-        Self(self.0 + rhs as u64)
-    }
-}
-
-impl Sub<Amount> for Amount {
-    type Output = Amount;
-
-    fn sub(self, rhs: Amount) -> Self::Output {
-        Self(self.0.saturating_sub(rhs.0))
-    }
-}
-
-impl Sub<u64> for Amount {
-    type Output = Amount;
-
-    fn sub(self, rhs: u64) -> Self::Output {
-        self - Self(rhs)
-    }
-}
-
-// converisons
-
-impl From<u64> for Amount {
-    fn from(value: u64) -> Self {
-        Amount(value)
-    }
-}
-
-impl std::str::FromStr for Amount {
-    type Err = anyhow::Error;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        s.parse::<Decimal>()
-            .map(|amt| Self((amt * MINA_SCALE_DEC).to_u64().expect("currency amount")))
-            .map_err(|e| anyhow!("{e}"))
-    }
-}
-
-// display
-
-impl std::fmt::Display for Amount {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", nanomina_to_mina(self.0))
-    }
-}
