+++ rust/src/ledger/token/mod.rs
@@ -1,9 +1,9 @@
 pub mod symbol;
 
-use super::{amount::Amount, public_key::PublicKey};
+use super::public_key::PublicKey;
 use crate::{
-    constants::MINA_TOKEN_ADDRESS, protocol::serialization_types::version_bytes::TOKEN_ID_KEY,
-    utility::store::common::U64_LEN,
+    base::amount::Amount, constants::MINA_TOKEN_ADDRESS,
+    protocol::serialization_types::version_bytes::TOKEN_ID_KEY, utility::store::common::U64_LEN,
 };
 use anyhow::bail;
 use serde::{Deserialize, Serialize};
