+++ rust/src/web/rest/blockchain.rs
@@ -1,9 +1,10 @@
 use crate::{
+    base::amount::Amount,
     block::{precomputed::PrecomputedBlock, store::BlockStore},
     chain::store::ChainStore,
     command::{internal::store::InternalCommandStore, store::UserCommandStore},
     constants::{MAINNET_EPOCH_SLOT_COUNT, VERSION},
-    ledger::{amount::Amount, store::best::BestLedgerStore},
+    ledger::store::best::BestLedgerStore,
     snark_work::store::SnarkStore,
     store::{
         version::{IndexerStoreVersion, VersionStore},
