+++ rust/tests/state/dangling_branches/add_all_blocks.rs
@@ -1,4 +1,5 @@
 use mina_indexer::{
+    base::amount::Amount,
     block::{parser::BlockParser, precomputed::PrecomputedBlock},
     constants::MINA_SCALE,
     ledger::{
